+++ trunk/xlib/xapp.h
@@ -74,7 +74,7 @@ public:
   The resulting map contains Levels, 0 for the surface, 1 - one pixel off the survface etc.
   returns the maximum level reached.  The void center is assigned to the point coordinates with
   largest level*/
-  short CalcVoid(TArray3D<short>& map, double extraR, short val, long* structurePoints, 
+  short CalcVoid(TArray3D<short>& map, double extraR, short val, size_t* structurePoints, 
     vec3d& voidCenter, TPSTypeList<TBasicAtomInfo*, double>* radii);
   /* function undoes renaming atoms */
   void undoName(TUndoData *data);
